backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
python-dotenv==1.0.1
httpx==0.27.0
beautifulsoup4==4.12.3
requests==2.31.0
langchain==0.1.20
langchain-google-genai==1.0.3
google-generativeai==0.5.4
pydantic==2.7.1
python-multipart==0.0.9
aiohttp==3.9.5
lxml==5.2.2
cloudscraper==1.2.71
json-repair==0.30.3
